// src/app/models/maintenance-plan.model.ts
export interface MaintenancePlan {
  id: String;
  worker: String;
  workid: String;
  periodicity: String;
  date: String;
  starttime: String;
  finishtime: String;
  totaltime: String;
  operationtime: String;
  replacementparts: String;
  ppe: String;
  risks: String;
  assetcode: String;
  task: String;
  taskrange: String;
  result: String;
  observations: String;
}